+++ crates/nostr/src/event/builder.rs
@@ -141,7 +141,7 @@ pub struct EventBuilder {
     // These if changed may break the previously constructed events
     // (i.e., change the content of a NIP46 event or of a NIP59 seal).
     kind: Kind,
-    tags: Vec<Tag>,
+    tags: Tags,
     content: String,
     /// Custom timestamp
     pub custom_created_at: Option<Timestamp>,
@@ -151,6 +151,10 @@ pub struct EventBuilder {
     ///
     /// If enabled, the `p` tags that match the signing keypair will not be discarded.
     pub allow_self_tagging: bool,
+    /// Deduplicate tags
+    ///
+    /// For more details check [`Tags::dedup`].
+    pub dedup_tags: bool,
 }
 
 impl EventBuilder {
@@ -162,11 +166,12 @@ impl EventBuilder {
     {
         Self {
             kind,
-            tags: Vec::new(),
+            tags: Tags::new(),
             content: content.into(),
             custom_created_at: None,
             pow: None,
             allow_self_tagging: false,
+            dedup_tags: false,
         }
     }
 
@@ -215,20 +220,26 @@ impl EventBuilder {
         self
     }
 
+    /// Deduplicate tags
+    ///
+    /// For more details check [`Tags::dedup`].
+    pub fn dedup_tags(mut self) -> Self {
+        self.dedup_tags = true;
+        self
+    }
+
     /// Build an unsigned event
     ///
     /// By default, this method removes any `p` tags that match the author's public key.
     /// To allow self-tagging, call [`EventBuilder::allow_self_tagging`] first.
-    pub fn build_with_ctx<T>(self, supplier: &T, public_key: PublicKey) -> UnsignedEvent
+    pub fn build_with_ctx<T>(mut self, supplier: &T, public_key: PublicKey) -> UnsignedEvent
     where
         T: TimeSupplier,
     {
-        let mut tags: Vec<Tag> = self.tags;
-
         // If self-tagging isn't allowed, discard all `p` tags that match the event author.
         if !self.allow_self_tagging {
             let public_key_hex: String = public_key.to_hex();
-            tags.retain(|t| {
+            self.tags.retain(|t| {
                 if t.kind() == TagKind::p() {
                     if let Some(content) = t.content() {
                         if content == public_key_hex {
@@ -241,23 +252,31 @@ impl EventBuilder {
             });
         }
 
+        // Deduplicate tags
+        if self.dedup_tags {
+            self.tags.dedup();
+        }
+
         // Check if should be POW
         match self.pow {
             Some(difficulty) if difficulty > 0 => {
                 let mut nonce: u128 = 0;
 
-                tags.reserve_exact(1);
-
                 loop {
                     nonce += 1;
 
-                    tags.push(Tag::pow(nonce, difficulty));
+                    self.tags.push(Tag::pow(nonce, difficulty));
 
                     let created_at: Timestamp = self
                         .custom_created_at
                         .unwrap_or_else(|| Timestamp::now_with_supplier(supplier));
-                    let id: EventId =
-                        EventId::new(&public_key, &created_at, &self.kind, &tags, &self.content);
+                    let id: EventId = EventId::new(
+                        &public_key,
+                        &created_at,
+                        &self.kind,
+                        &self.tags,
+                        &self.content,
+                    );
 
                     if id.check_pow(difficulty) {
                         return UnsignedEvent {
@@ -265,12 +284,12 @@ impl EventBuilder {
                             pubkey: public_key,
                             created_at,
                             kind: self.kind,
-                            tags: Tags::from_list(tags),
+                            tags: self.tags,
                             content: self.content,
                         };
                     }
 
-                    tags.pop();
+                    self.tags.pop();
                 }
             }
             // No POW difficulty set OR difficulty == 0
@@ -282,7 +301,7 @@ impl EventBuilder {
                         .custom_created_at
                         .unwrap_or_else(|| Timestamp::now_with_supplier(supplier)),
                     kind: self.kind,
-                    tags: Tags::from_list(tags),
+                    tags: self.tags,
                     content: self.content,
                 };
                 unsigned.ensure_id();
@@ -480,11 +499,8 @@ impl EventBuilder {
                 .cloned(),
         );
 
-        // Dedup tags
-        tags.dedup();
-
         // Compose event
-        Self::new(Kind::TextNote, content).tags(tags)
+        Self::new(Kind::TextNote, content).tags(tags).dedup_tags()
     }
 
     /// Comment
@@ -576,11 +592,8 @@ impl EventBuilder {
         // Add others `p` tags of comment_to event
         extend_nip22_p_tags(comment_to, &mut tags);
 
-        // Dedup tags
-        tags.dedup();
-
         // Compose event
-        Self::new(Kind::Comment, content).tags(tags)
+        Self::new(Kind::Comment, content).tags(tags).dedup_tags()
     }
 
     /// Long-form text note (generally referred to as "articles" or "blog posts").